(#47) Military status communicated to pybot during signup info flow
This adds a new background task to communicate to pybot.

When an update is performed, it checks that the slack id and military
status is present and not equal to old values and kicks off the
background task.

The test is not yet complete, it needs to exercise the background task.